Rename EmptySurfaceProcessor to CopyingSurfaceProcessor
